+++ src/arch/x64/structures/mod.rs
@@ -1,0 +1,6 @@
+// src/arch/x64/structures/mod.rs
+//
+// This is the mod.rs file for the arch::x64::struct module.
+
+
+pub mod port;